requirements.txt
Django~=5.1.3
Flask~=3.0.3
numpy~=1.26.4
pandas~=2.2.3
Requests~=2.32.2
yfinance~=0.2.51
django-cors-headers==4.3.0
psutil~=5.9.0
djangorestframework~=3.15.2
Pillow~=10.4.0